rename hook for valutakurs og -skjema og flytt til riktig mappe
