+++ src/frontend/sider/Fagsak/Behandling/sider/Behandlingsresultat/Eøs/Valutakurs/useValutakurs.ts
@@ -1,9 +1,8 @@
 import { useEffect, useState } from 'react';
 
-import { sorterEøsPerioder } from '../../sider/Fagsak/Behandling/sider/Behandlingsresultat/Eøs/utils';
-import type { IBehandling } from '../../typer/behandling';
-import type { IRestValutakurs } from '../../typer/eøsPerioder';
-import { EøsPeriodeStatus } from '../../typer/eøsPerioder';
+import type { IBehandling } from '../../../../../../../typer/behandling';
+import { EøsPeriodeStatus, type IRestValutakurs } from '../../../../../../../typer/eøsPerioder';
+import { sorterEøsPerioder } from '../utils';
 
 interface IProps {
     åpenBehandling: IBehandling;
